+++ src/errors/delete.rs
@@ -1,9 +1,12 @@
-use super::route::RouteError;
+use super::{route::RouteError, EncodingError};
 use std::{error::Error, fmt::Display};
 
 /// Errors relating to attempting to delete a route from a [`Router`](crate::Router).
 #[derive(Debug, PartialEq, Eq)]
 pub enum DeleteError {
+    /// A [`EncodingError`] that occurred during the decoding.
+    EncodingError(EncodingError),
+
     /// A [`RouteError`] that occurred during the delete.
     RouteError(RouteError),
 
@@ -39,6 +42,7 @@ impl Error for DeleteError {}
 impl Display for DeleteError {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
+            Self::EncodingError(error) => error.fmt(f),
             Self::RouteError(error) => error.fmt(f),
             Self::NotFound { route } => write!(
                 f,
@@ -49,6 +53,12 @@ impl Display for DeleteError {
 The specified route does not exist in the router"#
             ),
         }
+    }
+}
+
+impl From<EncodingError> for DeleteError {
+    fn from(error: EncodingError) -> Self {
+        Self::EncodingError(error)
     }
 }